refactored cards into seperate components
